+++ spotifywebapipython/models/showsimplified.py
@@ -184,10 +184,8 @@ class ShowSimplified:
         Gets the first image url in the `Images` list, if images are defined;
         otherwise, null.
         """
-        if len(self._Images) > 0:
-            return self._Images[0].Url
-        return None
-            
+        return ImageObject.GetImageHighestResolution(self._Images)
+    
         
     @property
     def IsExternallyHosted(self) -> bool: